refactored VsphereClient
